+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+
+- Fix resolving include commands from the `import` package ([#885](https://github.com/latex-lsp/texlab/issues/885))
+
 ## [5.5.1] - 2023-05-06
 
 ### Fixed